pyproject.toml
[project]
name = "propiedad_horizontal"
version = "0.1.0"
description = "Monolito FastAPI + Tortoise + SQLite + Poetry (layout src)"
authors = [
    {name = "Alexander Cifuentes Sanchez", email = "tu@email.com"}
    ]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "tortoise-orm (>=0.25.3,<0.26.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "debugpy (>=1.8.19,<2.0.0)",
    "qrcode (>=7.3,<8.0)",
    "httpx (>=0.28.1,<0.29.0)"
]

[tool.poetry]
packages = [{include = "propiedad_horizontal", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "aerich (>=0.9.2,<0.10.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]

[tool.aerich]
tortoise_orm = "propiedad_horizontal.app.core.db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./src"
